# docs/swagger.yaml
...
host: 127.0.0.1:8080
info:
  contact:
    email: fangbc5@163.com
    name: fangbc5
    url: http://chinaexam.top
  description: 应用程序API管理
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 考试系统接口文档
  version: v1.0.0
paths:
...
  /getLoginCode:
    get:
      consumes:
      - application/json
      description: 获取验证码
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 获取验证码
      tags:
      - 系统接口
  /getUserInfo:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 系统接口
  /login:
    post:
      consumes:
      - multipart/form-data
      description: 登录接口
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 验证码
        in: formData
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 登录接口
      tags:
      - 系统接口
  /logout:
    get:
      consumes:
      - application/json
      description: 登出接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 登出接口
      tags:
      - 系统接口
...
  /register:
    post:
      consumes:
      - application/json
      description: 注册接口
      parameters:
      - description: 1-学员，2-考官
        in: formData
        name: employerType
        required: true
        type: string
      - description: 名称/昵称
        in: formData
        name: employerName
        required: true
        type: string
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 注册接口
      tags:
      - 系统接口
swagger: "2.0"